requirements.txt
selenium
fake-useragent
google-cloud-bigquery